backend/requirements.txt
Django==4.2.*
djangorestframework==3.14.*
python-dotenv==1.0.*
openai==1.65.*
psycopg2-binary==2.9.*
django-cors-headers==3.14.*
tiktoken==0.9.*
pyyaml==6.0.*
arxiv==2.1.*
langchain-core==0.3.*
langgraph==0.3.*
langchain[openai]==0.3.*
pymupdf==1.25.*
jinja2==3.1.*